requirements.txt
biopython>=1.81
pandas>=2.0
requests>=2.31
pyyaml>=6.0
synapseclient>=4.0
